refactor: split device details view into modules
