+++ tankctl_app/lib/features/tank_detail/sections/device_info_section.dart
@@ -1,0 +1,71 @@
+import 'package:flutter/material.dart';
+import 'package:tankctl_app/domain/device_detail.dart';
+
+/// Device info section (read-only)
+class DeviceInfoSection extends StatelessWidget {
+  final DeviceDetail device;
+
+  const DeviceInfoSection({super.key, required this.device});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            'Device Info',
+            style: Theme.of(context).textTheme.titleMedium?.copyWith(
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+          const SizedBox(height: 12),
+          Container(
+            padding: const EdgeInsets.all(16),
+            decoration: BoxDecoration(
+              color: Colors.grey.withValues(alpha: 0.05),
+              borderRadius: BorderRadius.circular(12),
+              border: Border.all(color: Colors.grey.withValues(alpha: 0.2)),
+            ),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                _buildInfoRow(context, 'Device ID', device.deviceId),
+                const Divider(),
+                _buildInfoRow(context, 'Firmware', device.firmwareVersion ?? '—'),
+                const Divider(),
+                _buildInfoRow(context, 'Connection', device.status),
+                const Divider(),
+                _buildInfoRow(context, 'RSSI', '${device.rssi ?? 0} dBm'),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildInfoRow(BuildContext context, String label, String value) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text(
+            label,
+            style: Theme.of(context).textTheme.bodySmall?.copyWith(
+              color: Colors.grey,
+            ),
+          ),
+          Text(
+            value,
+            style: Theme.of(context).textTheme.bodySmall?.copyWith(
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
